# Conflicts:
#	src/main/java/Alumni/backend/module/controller/TeamController.java
#	src/main/java/Alumni/backend/module/domain/community/Comment.java

+++ src/main/java/Alumni/backend/module/service/community/CommentService.java
@@ -5,12 +5,14 @@ import Alumni.backend.infra.event.community.RecommentCreateEvent;
 import Alumni.backend.infra.exception.NoExistsException;
 import Alumni.backend.module.domain.community.Comment;
 import Alumni.backend.module.domain.community.CommentLike;
+import Alumni.backend.module.domain.contest.Team;
 import Alumni.backend.module.domain.registration.Member;
 import Alumni.backend.module.domain.community.Post;
 import Alumni.backend.module.repository.community.CommentLikeRepository;
 import java.util.List;
 import Alumni.backend.module.repository.community.comment.CommentRepository;
 import Alumni.backend.module.repository.community.post.PostRepository;
+import Alumni.backend.module.repository.contest.TeamRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
@@ -27,7 +29,11 @@ public class CommentService {
     private final CommentRepository commentRepository;
     private final CommentLikeRepository commentLikeRepository;
     private final ApplicationEventPublisher eventPublisher;
+    private final TeamRepository teamRepository;
 
+    /**
+     * 게시글 댓글 관련 메소드
+     */
     public Long createComment(Member member, Long postId, String content) {
         if (content.length() == 0)
             throw new IllegalArgumentException("Bad Request");
@@ -36,7 +42,8 @@ public class CommentService {
         if (post == null) {
             throw new NoExistsException("존재하지 않는 게시글입니다");
         }
-        Comment comment = Comment.createComment(member, post, content);
+        Comment comment = Comment.createComment(member, content);
+        comment.setPost(post);
         Comment saveComment = commentRepository.save(comment);
         postRepository.updateCommentCount(post.getCommentNum() + 1, postId);//게시글에 달린 댓글 수 증가
         // Async 알림 전송
@@ -44,19 +51,6 @@ public class CommentService {
         return saveComment.getId();
     }
 
-    public void modifyComment(Member member, Long commentId, String content) {
-
-        if (content.length() == 0)
-            throw new IllegalArgumentException("Bad Request");
-
-        Comment comment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new NoExistsException("존재하지 않는 댓글입니다"));
-
-        if (!comment.getMember().getId().equals(member.getId()))//수정하는 사람과 작성자가 같은지 확인
-            throw new IllegalArgumentException("Bad Request");
-
-        comment.modifyComment(content);
-    }
 
     public void deleteComment(Member member, Long commentId) {
         Comment comment = commentRepository.findById(commentId)
@@ -92,27 +86,13 @@ public class CommentService {
             throw new NoExistsException("상위 댓글이 존재하지 않습니다");
         }
         Post post = postRepository.findById(parent.getPost().getId()).orElseThrow(() -> new NoExistsException("존재하지 않는 게시글입니다"));
-        Comment recomment = Comment.createComment(member, post, content);
+        Comment recomment = Comment.createComment(member,content);
+        recomment.setPost(post);
         recomment.setParent(parent);
         commentRepository.save(recomment);
         postRepository.updateCommentCount(post.getCommentNum() + 1, parent.getPost().getId());//게시글에 달린 댓글 수 증가
         // Async 알림 전송
         eventPublisher.publishEvent(new RecommentCreateEvent(parent.getMember()));
-    }
-
-    public void modifyRecomment(Member member, Long commentId, String content) {
-        if (content.length() == 0)
-            throw new IllegalArgumentException("Bad Request");
-
-        Comment recomment = commentRepository.findById(commentId)
-                .orElseThrow(() -> new NoExistsException("존재하지 않는 대댓글입니다"));
-
-        Post post = postRepository.findById(recomment.getPost().getId()).orElseThrow(() -> new NoExistsException("존재하지 않는 게시글입니다"));
-
-        if (!recomment.getMember().getId().equals(member.getId()))//수정하는 사람과 작성자가 같은지 확인
-            throw new IllegalArgumentException("Bad Request");
-
-        recomment.modifyComment(content);
     }
 
     public void deleteRecomment(Member member, Long commentId) {
@@ -127,6 +107,51 @@ public class CommentService {
 
         List<CommentLike> recommentLikeList = commentLikeRepository.findByCommentId(recomment.getId());
         commentLikeRepository.deleteAll(recommentLikeList);
+
+        commentRepository.delete(recomment);
+    }
+
+    /**
+     * 공모전 팀 댓글 관련 메소드
+     */
+    public void createTeamComment(Member member, Long teamId, String content) {
+        if (content.length() == 0)
+            throw new IllegalArgumentException("Bad Request");
+
+        Team team = teamRepository.findById(teamId).orElseThrow(() -> new NoExistsException("존재하지 않는 팀"));
+
+        Comment teamComment = Comment.createComment(member, content);
+        teamComment.setTeam(team);
+
+        commentRepository.save(teamComment);
+    }
+
+    public void deleteTeamComment(Member member, Long commentId) {
+        Comment teamComment = commentRepository.findById(commentId).orElseThrow(() -> new NoExistsException("존재하지 않는 댓글"));
+
+        if(!member.getId().equals(teamComment.getMember().getId())) //동일한 사람인지 확인
+            throw new IllegalArgumentException("Bad Request");
+
+        commentRepository.delete(teamComment);
+    }
+
+    public void createTeamRecomment(Member member, Long commentId, String content) {
+        if (content.length() == 0)
+            throw new IllegalArgumentException("Bad Request");
+
+        Comment parentComment = commentRepository.findById(commentId).orElseThrow(() -> new NoExistsException("상위 댓글이 존재하지 않습니다"));
+        Comment recomment = Comment.createComment(member, content);
+        recomment.setTeam(parentComment.getTeam());
+        recomment.setParent(parentComment);
+
+        commentRepository.save(recomment);
+    }
+
+    public void deleteTeamRecomment(Member member, Long commentId) {
+        Comment recomment = commentRepository.findById(commentId).orElseThrow(() -> new NoExistsException("존재하지 않는 댓글"));
+
+        if(!member.getId().equals(recomment.getMember().getId())) //동일한 사람인지 확인
+            throw new IllegalArgumentException("Bad Request");
 
         commentRepository.delete(recomment);
     }
